move `CEditor::DoMapEditor` to map_view.cpp
